Add Authorization in Delete

+++ Part-4/controllers/blog.controller.js
@@ -42,6 +42,18 @@ const createBlogs = async (req, res, next) => {
 
 const removeBlog = async (req, res, next) => {
   try {
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    if (!decodedToken.id) {
+      return res.status(401).json({ error: "token invalid" });
+    }
+    const blog = await Blog.findById(req.params.id);
+    if (!blog) {
+      return res.status(404).json({ error: "blog not found" });
+    }
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return res.status(403).json({ error: "only creator can delete blog" });
+    }
     await Blog.findByIdAndDelete(req.params.id);
     res.status(204).end();
   } catch (error) {
